refactor(eve): split Vercel sandbox session adapter
